UI_Reconstruct
Reconstruct Hierarchy
