--- UI_Address/MainApp.java
@@ -1,70 +1,0 @@
-package UI_Address;
-
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.BorderPane;
-import javafx.stage.Stage;
-
-public class MainApp extends Application {
-
-    private Stage primaryStage;
-    //private BorderPane rootLayout;
-
-    @Override
-    public void start(Stage primaryStage) {
-        this.primaryStage = primaryStage;
-        this.primaryStage.setTitle("AddressApp");
-
-        //initRootLayout();
-
-        showMain();
-    }
-
-    /**
-    public void initRootLayout() {
-        try {
-            // fxml 파일에서 상위 레이아웃을 가져온다.
-            FXMLLoader loader = new FXMLLoader();
-            loader.setLocation(MainApp.class.getResource("view/RootLayout.fxml"));
-            rootLayout = (BorderPane) loader.load();
-
-            // 상위 레이아웃을 포함하는 scene을 보여준다.
-            Scene scene = new Scene(rootLayout);
-            primaryStage.setScene(scene);
-            primaryStage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-     */
-    public void showMain() {
-        try {
-            FXMLLoader loader = new FXMLLoader();
-            loader.setLocation(MainApp.class.getResource("view/Main.fxml"));
-            AnchorPane main = (AnchorPane) loader.load();
-            
-            Scene scene = new Scene(main);
-            primaryStage.setScene(scene);
-            primaryStage.show();
-            // 연락처 요약을 상위 레이아웃 가운데로 설정한다.
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * 메인 스테이지를 반환한다.
-     * @return
-     */
-    public Stage getPrimaryStage() {
-        return primaryStage;
-    }
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-}
