+++ src/main/java/UI/Controller.java
@@ -1,0 +1,4 @@
+package sample;
+
+public class Controller {
+}
