+++ src/main/java/UI/Main.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<AnchorPane prefHeight="478.0" prefWidth="731.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+   <children>
+      <VBox prefHeight="478.0" prefWidth="132.0" styleClass="sidebar" stylesheets="@ui.css">
+         <children>
+            <HBox prefHeight="50.0" prefWidth="132.0" />
+         </children>
+      </VBox>
+   </children>
+</AnchorPane>
