+++ src/main/java/UI/ui.css
@@ -1,0 +1,6 @@
+.parent{
+	-fx-background-color: #fff;
+}
+.sidebar{
+	-fx-background-color: #333645
+}
