--- src/main/java/UI_Address/.gitignore
@@ -1,1 +1,0 @@
-/MainApp.class
