--- src/main/java/UI_Address/view/Main.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane prefHeight="500.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Pane layoutX="171.0" layoutY="164.0" prefHeight="500.0" prefWidth="500.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-         <children>
-            <Button layoutX="223.0" layoutY="239.0" mnemonicParsing="false" text="Button" />
-            <ListView layoutX="277.0" layoutY="27.0" prefHeight="200.0" prefWidth="200.0" />
-            <ProgressBar layoutX="286.0" layoutY="242.0" prefWidth="200.0" progress="0.0" />
-         </children>
-      </Pane>
-   </children>
-</AnchorPane>
